src/database/requirements.txt
supabase==1.0.3
pandas==2.1.0
matplotlib==3.7.1
seaborn==0.12.2
streamlit==1.24.0
numpy==1.24.3
python-dotenv==1.0.0
plotly==5.18.0
scipy==1.11.3
statsmodels==0.14.0 